chore: move theme decorator
